refactor: Extract extractZip function
